GCOMB/requirements.txt
tqdm==4.57.0
networkx==1.11
pandas==1.1.5
tensorflow-gpu==1.13.1
tensorboard==1.8.0
scipy==1.1.0
numpy==1.19.5
scikit-learn==0.19.1